feat: commands now store / command data
tweaked the command and trigger storage
default all trigger strings will append /say to the beginning. ONLY if the string does not start with a /
this makes piecing together eqbc and dannet strings easier.

This also allows you to store other commands.
say i target a merchant and want to set a string to execute /lootutils sellstuff

adds a little more versatility

this will auto update your existing tables. but shouldn't impact any usage. Recommend testing with debug on to see the outputs.